# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 이룸 (Yiroom) API
  description: |
    온전한 나를 위한 통합 웰니스 AI 플랫폼 API

    ## 개요
    이룸은 AI 기반 뷰티/헬스 분석 및 개인화 추천 서비스를 제공합니다.
    - 퍼스널컬러 분석 (12톤 시스템)
    - 피부 분석 (7존 분석)
    - 체형 분석 (MediaPipe 33 랜드마크)
    - 헤어 분석 (얼굴형 기반)
    - 구강건강 분석 (VITA 셰이드)
    - AI 웰니스 코치

    ## 인증
    모든 API는 Clerk 인증이 필요합니다. Authorization 헤더에 Bearer 토큰을 포함하세요.
    ```
    Authorization: Bearer <clerk_jwt_token>
    ```

    ## Rate Limiting
    - 분석 API: 50 requests / 24h / user
    - 일반 API: 100 requests / 1min / user
    - Rate Limit 초과 시 429 응답과 함께 `Retry-After` 헤더 반환

    ## Mock Fallback
    모든 AI 분석 API는 3초 타임아웃 + 2회 재시도 후 Mock 데이터로 폴백합니다.
    응답의 `usedFallback` 필드로 확인 가능합니다.

    ## 에러 응답 형식
    ```json
    {
      "error": "사용자 친화적 에러 메시지",
      "code": "ERROR_CODE",
      "details": "상세 정보 (개발 환경에서만)",
      "retryAfter": 3600
    }
    ```

    ## 에러 코드
    | 코드 | HTTP | 설명 |
    |------|------|------|
    | UNAUTHORIZED | 401 | 인증 필요 |
    | FORBIDDEN | 403 | 권한 없음 |
    | VALIDATION_ERROR | 400 | 입력값 검증 실패 |
    | NOT_FOUND | 404 | 리소스 없음 |
    | RATE_LIMIT_EXCEEDED | 429 | 요청 제한 초과 |
    | IMAGE_QUALITY_ERROR | 422 | 이미지 품질 부적합 |
    | AI_SERVICE_ERROR | 500 | AI 서비스 오류 |
    | DB_ERROR | 500 | 데이터베이스 오류 |
    | INTERNAL_ERROR | 500 | 서버 오류 |
  version: 2.1.0
  contact:
    name: Yiroom Team
    url: https://yiroom.app
    email: dev@yiroom.app
  license:
    name: Proprietary
servers:
  - url: https://yiroom.app/api
    description: Production
...
paths:
...
  /workout/stretching:
    post:
      tags: [Workout]
      summary: 스트레칭 루틴 생성
      description: 자세 분석 결과 기반 맞춤 스트레칭 루틴
      operationId: generateStretchingRoutine
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                postureIssues:
                  type: array
                  items:
                    type: string
                duration:
                  type: integer
                  default: 15
      responses:
        '200':
          description: 생성 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - ClerkAuth: []

  # ============================================================================
  # Coach API
  # ============================================================================
...
  /smart-matching/size-recommend:
    post:
      tags: [Products]
      summary: 사이즈 추천
      description: 체형 분석 기반 의류 사이즈 추천
      operationId: recommendSize
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                brand:
                  type: string
                category:
                  type: string
                  enum: [top, bottom, dress, outer]
                measurements:
                  $ref: '#/components/schemas/BodyMeasurements'
      responses:
        '200':
          description: 추천 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SizeRecommendation'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - ClerkAuth: []
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT Token
  responses:
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTH_ERROR
              message: User not authenticated
              userMessage: 로그인이 필요합니다.
...
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message, userMessage]
          properties:
            code:
              type: string
              enum:
                - AUTH_ERROR
                - VALIDATION_ERROR
                - RATE_LIMIT_ERROR
                - AI_TIMEOUT_ERROR
                - DB_ERROR
                - INTERNAL_ERROR
            message:
              type: string
            userMessage:
              type: string
            details:
              type: object
...
    BodyMeasurements:
      type: object
      properties:
        height:
          type: number
          description: 키 (cm)
        weight:
          type: number
          description: 체중 (kg)
        chest:
          type: number
          description: 가슴둘레 (cm)
        waist:
          type: number
          description: 허리둘레 (cm)
        hip:
          type: number
          description: 엉덩이둘레 (cm)
    SizeRecommendation:
      type: object
      properties:
        recommendedSize:
          type: string
          enum: [XS, S, M, L, XL, XXL]
        confidence:
          type: number
        alternatives:
          type: array
          items:
            type: object
            properties:
              size:
                type: string
              fitType:
                type: string
                enum: [slim, regular, relaxed]
...